move to using require('webpack').container.ModuleFedreationPlugin
