+++ bi-directional/app2/webpack.config.js
@@ -1,5 +1,6 @@
 const HtmlWebpackPlugin = require("html-webpack-plugin");
-const ModuleFederationPlugin = require("webpack/lib/container/ModuleFederationPlugin");
+const ModuleFederationPlugin = require("webpack").container
+  .ModuleFederationPlugin;
 const path = require("path");
 const sharedReduce = ["react", "react-dom"].reduce((shared, pkg) => {
   Object.assign(shared, { [`${pkg}-${require(pkg).version}`]: pkg });
